+++ theme/template/_includes/social.php
@@ -1,4 +1,8 @@
 <section>
-  <a class="icon icon--twitter" href="#" target="_blank"></a>
-  <a class="icon icon--facebook" href="#" target="_blank"></a>
+  <a class="icon icon--twitter" href="<?= tweet_url(
+    $post
+  ) ?>" target="_blank"></a>
+  <a class="icon icon--facebook" href="<?= facebook_url(
+    $post
+  ) ?>" target="_blank"></a>
 </section>